% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipe.R
\name{get_recipe}
\alias{get_recipe}
\title{scrape the details of each recipe from NYTCooking website}
\usage{
get_recipe(url)
}
\arguments{
\item{url}{a url linking to a recipe from Nytimes cooking website}
}
\description{
scrape the details of each recipe from NYTCooking website
}
\examples{
get_recipe("https://cooking.nytimes.com/recipes/1022411-cucumber-avocado-salad")
}
