% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_item.R
\name{match_item}
\alias{match_item}
\title{match user input items with ingredients from recipes}
\usage{
match_item(...)
}
\arguments{
\item{...}{}
}
\value{
10 recipes based on the ingredient import
}
\description{
match user input items with ingredients from recipes
}
\examples{
test<-match_item("apple","egg","sugar","vanilla")
}
