% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ingredients.R
\name{get_ingredients}
\alias{get_ingredients}
\title{get ingredient list from user and return matched recipes as output}
\usage{
get_ingredients(...)
}
\arguments{
\item{...}{because the function directly asks for user input, the argument is not necessary, but it can be a vector of character that contains ingredient names}
}
\value{
a 10-row dataframe of recipes based on user input
}
\description{
get ingredient list from user and return matched recipes as output
}
